Organize file download ownership
