+++ mobile/apps/photos/lib/module/upload/service/local_file_update_service.dart
@@ -10,9 +10,9 @@ import 'package:photos/db/file_updation_db.dart';
 import 'package:photos/db/files_db.dart';
 import 'package:photos/models/file/file.dart';
 import 'package:photos/models/file/file_type.dart';
+import 'package:photos/module/download/file.dart';
 import 'package:photos/module/upload/model/media_upload_data.dart';
 import "package:photos/module/upload/upload_data.dart";
-import 'package:photos/utils/file_util.dart';
 import 'package:shared_preferences/shared_preferences.dart';
 
 /// Finds device files whose contents changed after upload and queues updates.
