+++ mobile/apps/photos/lib/services/video_memory_service.dart
@@ -11,9 +11,9 @@ import "package:path_provider/path_provider.dart";
 import "package:photos/generated/l10n.dart";
 import "package:photos/models/file/file.dart";
 import "package:photos/models/file/file_type.dart";
+import "package:photos/module/download/file.dart";
 import "package:photos/ui/notification/toast.dart";
 import "package:photos/utils/dialog_util.dart";
-import "package:photos/utils/file_util.dart";
 
 final _logger = Logger("VideoMemoryService");
