+++ mobile/apps/photos/lib/services/video_preview_service.dart
@@ -33,6 +33,7 @@ import "package:photos/models/metadata/file_magic.dart";
 import "package:photos/models/preview/playlist_data.dart";
 import "package:photos/models/preview/preview_item.dart";
 import "package:photos/models/preview/preview_item_status.dart";
+import "package:photos/module/download/file.dart";
 import "package:photos/module/metadata/video.dart";
 import "package:photos/module/upload/service/file_uploader.dart";
 import "package:photos/service_locator.dart";
@@ -42,7 +43,6 @@ import "package:photos/services/isolated_ffmpeg_service.dart";
 import "package:photos/services/machine_learning/compute_controller.dart";
 import "package:photos/ui/notification/toast.dart";
 import "package:photos/utils/file_key.dart";
-import "package:photos/utils/file_util.dart";
 import "package:photos/utils/gzip.dart";
 import "package:photos/utils/network_util.dart";
