+++ mobile/apps/photos/lib/ui/viewer/file/inline_text_detection.dart
@@ -14,13 +14,13 @@ import "package:photos/l10n/l10n.dart";
 import "package:photos/models/file/file.dart";
 import "package:photos/models/file/file_type.dart";
 import "package:photos/models/file/trash_file.dart";
+import "package:photos/module/download/file.dart";
 import "package:photos/states/detail_page_state.dart";
 import "package:photos/ui/viewer/file/ocr/display_image_helper.dart";
 import "package:photos/ui/viewer/file/ocr/ocr_dot_wave_overlay.dart";
 import "package:photos/ui/viewer/file/ocr/text_detector_widget.dart";
 import "package:photos/ui/viewer/file/ocr/text_overlay_widget.dart"
     show ZoomedInteractionPolicy;
-import "package:photos/utils/file_util.dart";
 import "package:photos/utils/image_util.dart";
 
 /// Inline text detection widget that mimics Apple's Live Text behavior:
