+++ mobile/apps/photos/lib/ui/viewer/file/qr_code_detection_helper.dart
@@ -6,7 +6,7 @@ import "package:flutter/foundation.dart";
 import "package:logging/logging.dart";
 import "package:photos/models/file/file.dart";
 import "package:photos/models/file/file_type.dart";
-import "package:photos/utils/file_util.dart";
+import "package:photos/module/download/file.dart";
 
 class QrCodeDetectionHelper {
   static const _debounceDuration = Duration(milliseconds: 500);
