+++ mobile/apps/photos/lib/ui/viewer/file/zoomable_live_image_new.dart
@@ -13,12 +13,12 @@ import "package:photos/generated/l10n.dart";
 import "package:photos/models/file/extensions/file_props.dart";
 import 'package:photos/models/file/file.dart';
 import "package:photos/models/metadata/file_magic.dart";
+import 'package:photos/module/download/file.dart';
 import "package:photos/services/file_magic_service.dart";
 import "package:photos/src/rust/api/motion_photo_api.dart";
 import "package:photos/states/detail_page_state.dart";
 import 'package:photos/ui/notification/toast.dart';
 import 'package:photos/ui/viewer/file/zoomable_image.dart';
-import 'package:photos/utils/file_util.dart';
 
 class ZoomableLiveImageNew extends StatefulWidget {
   final EnteFile enteFile;
