+++ mobile/apps/photos/lib/ui/viewer/file_details/file_properties_item_widget.dart
@@ -5,9 +5,9 @@ import 'package:path/path.dart' as path;
 import "package:photos/models/file/extensions/file_props.dart";
 import 'package:photos/models/file/file.dart';
 import 'package:photos/models/file/file_type.dart';
+import "package:photos/module/download/file.dart";
 import "package:photos/theme/ente_theme.dart";
 import "package:photos/ui/components/info_item_widget.dart";
-import "package:photos/utils/file_util.dart";
 import "package:photos/utils/image_util.dart";
 import "package:photos/utils/magic_util.dart";
