+++ mobile/apps/photos/lib/utils/gallery_save_title.dart
@@ -1,7 +1,12 @@
+import "dart:io";
+
 import "package:photos/core/constants.dart";
 import "package:photos/utils/device_info.dart";
 
 Future<String> getMediaStoreCompatibleTitle(String title) async {
+  if (!Platform.isAndroid) {
+    return title;
+  }
   final sanitizedTitle = _sanitizePreAndroid11MediaStoreTitle(title);
   if (sanitizedTitle == title) {
     return title;
