+++ mobile/apps/photos/lib/utils/share_util.dart
@@ -10,10 +10,10 @@ import 'package:photos/core/configuration.dart';
 import 'package:photos/core/constants.dart';
 import 'package:photos/models/file/file.dart';
 import 'package:photos/models/file/file_type.dart';
+import 'package:photos/module/download/file.dart';
 import 'package:photos/module/metadata/exif.dart';
 import 'package:photos/module/metadata/local_file.dart';
 import 'package:photos/utils/dialog_util.dart';
-import 'package:photos/utils/file_util.dart';
 import 'package:receive_sharing_intent/receive_sharing_intent.dart';
 import 'package:share_plus/share_plus.dart';
 import "package:uuid/uuid.dart";
